+++ docs/source/ttnn/ttnn/ttnn/erf_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.erf_bw:
+
+ttnn.erf_bw
+###############
+
+.. autofunction:: ttnn.erf_bw
