+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -133,7 +133,7 @@ std::vector<Tensor> polygamma_bw(
     const Tensor& input,
     int n,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-    
+
 std::vector<Tensor> hardtanh_bw(
     const Tensor& grad,
     const Tensor& input,
@@ -145,11 +145,6 @@ std::vector<Tensor> angle_bw(
     const Tensor& grad,
     const Tensor& input,
     bool is_complextensor = true,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
-std::vector<Tensor> erf_bw(
-    const Tensor& grad,
-    const Tensor& input,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
 std::vector<Tensor> deg2rad_bw(
