+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -84,6 +84,7 @@ enum class UnaryBackwardOpType {
     RECIPROCAL_BW,
     DIGAMMA_BW,
     ERFINV_BW,
+    ERF_BW,
 };
 
 struct UnaryBackwardFunction{
